backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
langchain-openai
langchain-text-splitters
langgraph
langsmith
openai
sentence-transformers
supabase
python-dotenv
